More refinements

+++ ConnectServo.cpp
@@ -5,7 +5,7 @@ ServoQueueItem::ServoQueueItem() {
 
 };
 
-void ServoQueueItem::assign(uint8_t newCall, uint8_t newParam1, uint8_t newAnimationType, uint8_t newServoSpeed) {
+void ServoQueueItem::assign(uint8_t newCall, uint8_t newParam1, uint8_t newAnimationType, uint16_t newServoSpeed) {
     call = newCall;
     param1 = newParam1;
     animationType = newAnimationType;
